pyproject.toml
[project]
name = "oil-spill-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.126.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
